+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.1] - ?
+### Changed
+ - Added better support for dealing with exceptions thrown when clearing
+   threadlocals with `ThreadLocalSigner`. Added relevant exception class:
+   `ThreadLocalClearException`.
+
 ## [2.0.0] - 2016-01-07
 ### Changed
  - Renamed HttpSigner to Signer and changed it from a static utility class